refactor split components in admin posts page into smaller files for better maintainability, remove unused AdminHeader component
