+++ apps/web/app/(dashboard)/admin/posts/_components/PostDetailModal/utils.ts
@@ -1,0 +1,17 @@
+export const formatPrice = (price: string | number): string => {
+  const numPrice = typeof price === 'string' ? parseFloat(price) : price;
+  return new Intl.NumberFormat('vi-VN', {
+    style: 'currency',
+    currency: 'VND',
+  }).format(numPrice);
+};
+
+export const formatDateTime = (dateString: string): string => {
+  return new Date(dateString).toLocaleString('vi-VN', {
+    hour: '2-digit',
+    minute: '2-digit',
+    day: '2-digit',
+    month: '2-digit',
+    year: 'numeric',
+  });
+};
