+++ apps/web/app/(dashboard)/admin/posts/_components/index.ts
@@ -1,4 +1,3 @@
-export { AdminHeader } from './AdminHeader';
 export { StatusSummaryCards } from './StatusSummaryCards';
 export { FilterButtons } from './FilterButtons';
 export { PostCard } from './PostCard';
